Cleanup a precision-loss warning in GUS
Warning is:

  Implicit conversion loses integer precision: 'const int'
  to 'const uint16_t' (aka 'const unsigned short')
      const uint16_t frames = std::min(BUFFER_FRAMES,
                              requested_frames - generated_frames)
            ~~~~~~   ^~~~~~~~~~~~~~~~~~~~~~~

We can safety store the result of requested_frames - generated_frames
in a uint16_t because:

 1) Requested_frames is already bound to the uint16_t range both from
    it's inbound type and the mixer's callback API, which operates on
    at-most uint16_t frames.

 2) This calculation will never be negative because generated_frames
    is always guaranteed to be less-than the requested_frames given
    the loop conditional: while (generated_frames < requested_frames)
    therefore, there is no risk of underflow.

+++ src/hardware/gus.cpp
@@ -649,8 +649,9 @@ void Gus::AudioCallback(const uint16_t requested_frames)
 		for (auto &val : accumulator)
 			val = 0;
 
-		const uint16_t frames = std::min(BUFFER_FRAMES,
-		                                 requested_frames - generated_frames);
+		const uint16_t frames = static_cast<uint16_t>(
+		        std::min(BUFFER_FRAMES, requested_frames - generated_frames));
+
 		auto v = voices.begin();
 		const auto v_end = v + active_voices;
 		while (v < v_end && *v) {
